requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
flake8==6.0.0
flake8-black==0.3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
